+++ src/main/java/com/bullimog/portal/models/Temperatures.java
@@ -4,24 +4,24 @@ import java.time.LocalDateTime;
 
 public class Temperatures {
     private  ArrayList<LocalDateTime> dateTimes;
-    private  ArrayList<Float> temperatures;
+    private  ArrayList<Double> temperatures;
 
     public Temperatures(){
         dateTimes = new ArrayList<LocalDateTime>();
-        temperatures = new ArrayList<Float>();
+        temperatures = new ArrayList<Double>();
     }
 
-    public Temperatures(ArrayList<LocalDateTime> dateTimes, ArrayList<Float> temperatures){
+    public Temperatures(ArrayList<LocalDateTime> dateTimes, ArrayList<Double> temperatures){
         this.dateTimes = dateTimes;
         this.temperatures = temperatures;
     }
 
-    public void appendTemperature(Float temperature){
+    public void appendTemperature(Double temperature){
         LocalDateTime date = LocalDateTime.now();
         dateTimes.add(date);
         temperatures.add(temperature);
     }
 
     public ArrayList<LocalDateTime> getdateTime() {return dateTimes;}
-    public ArrayList<Float> getTemperature(){return temperatures;}
+    public ArrayList<Double> getTemperature(){return temperatures;}
 }
